More registers and settings.
I did a bit more research into how the I2C communication is supposed to
work. I *think* we are supposed to acutally set the START bit when
starting an I2C transaction, which I added. Also, added more register
settings related to routing of the DAC to the proper amplifiers. Not
ready for master.